Feature/sync aura price from geckoterminal (#487)
(cherry picked from commit db41715f23904476b447fb97ba35509789e9afd1)
