+++ .env-example
@@ -50,3 +50,6 @@ SMART_CONTRACT_SERVICE=
 INDEXER_V2_URL=https://indexer-v2.dev.aurascan.io/
 INDEXER_V2_DB=
 INDEXER_V2_SECRET=
+
+# geckoterminal api
+GECKOTERMINAL_API=https://api.geckoterminal.com/api/v2/
